Fixing the fixed map's seeding issue
The fixed map is supposed to remain fixed during the training even between parallel procs. However in training_agent.py we both set different seeds and generated the envs sequentially which resulted in each process having its own version of the map. In this change we remove seeding the parallel envs and we "reset" the seed for the fixed environment  each time. This way all parallel runs work on the same fixed map.

Notice that this does not break the non-fixed case because event hough are are seeded the same way from the training_agent.py, the envs are created sequentially. This results in sequential calls to the random number generator causing random maps.