First version of the evaluator logic
Added an evaluator.py file that gets called frequently from within the train_agent.py to evaluate the most recent model on a validation dataset. The validation dataset can be sampled form a dedicated LTLSampler or we can use the same sampler as the one for training.

For now this class is called from the same process as the trainer, perhaps later can move it to it's own dedicated process.