Refactoring the ltl_wrappers and polishing the SafetyEnv
Removed the LTLLetterEnv and made LTLEnv the main wrapper that adds the LTL logic to the envs.

On another side, renamed the SafetyEnv to ZonesEnv. LTLZonesEnv adds the LTL functions (get_propositions and get_events). Finally PlayEnv is now a wrapper that exposes key press events as part of the observations.